requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.0.0
psutil>=5.9.0
gputil>=1.4.0
sentencepiece>=0.1.99
protobuf>=3.20.0 
datasets>=2.14.0